backend/pyproject.toml
[tool.poetry]
name = "multimodal-chatbot"
version = "1.0.0"
description = "AI-powered multimodal chatbot with document processing"
authors = ["Your Name <email@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.104.1"
uvicorn = {extras = ["standard"], version = "^0.24.0"}
pydantic = "^2.5.2"
pydantic-settings = "^2.1.0"
sqlalchemy = "^2.0.23"
aiosqlite = "^0.19.0"
alembic = "^1.13.1"
pypdf = "^3.17.4"
python-docx = "^1.1.0"
pillow = "^10.1.0"
pytesseract = "^0.3.10"
markdown = "^3.5.1"
beautifulsoup4 = "^4.12.2"
chardet = "^5.2.0"
langchain = "^0.1.0"
langchain-community = "^0.0.10"
sentence-transformers = "^2.2.2"
chromadb = "^0.4.22"
httpx = "^0.25.2"
tenacity = "^8.2.3"
loguru = "^0.7.2"
python-multipart = "^0.0.6"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-dotenv = "^1.0.0"
aiofiles = "^23.2.1"
redis = "^5.0.1"
celery = "^5.3.4"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
black = "^23.12.1"
isort = "^5.13.2"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true